requirements.txt
flask
slack-bolt
langchain
langchain-core
langchain-groq
sqlalchemy
psycopg2-binary
python-dotenv